+++ backend/src/main/java/codesquad/team4/issuetracker/issue/dto/IssueRequestDto.java
@@ -1,5 +1,6 @@
 package codesquad.team4.issuetracker.issue.dto;
 
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,8 +10,12 @@ public class IssueRequestDto {
     @AllArgsConstructor
     @Getter
     @Builder
-    public static class CreateIssueDto {
-        private Long id;
-        private String messqge;
+    public static class CreateIssueDto{
+        private String title;
+        private String content;
+        private Long authorId;
+        private List<Long> assigneeId;
+        private List<Long> labelId;
+        private Long milestoneId;
     }
 }
